% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_CIs.R
\name{boot_coxrfx}
\alias{boot_coxrfx}
\title{Bootstrap confidence intervals for regression coefficients}
\usage{
boot_coxrfx(
  mstate_data_expanded,
  which_group,
  min_nr_samples = 100,
  output = "CIs",
  ...
)
}
\arguments{
\item{mstate_data_expanded}{Data in `long format`, possibly with `expanded` covariates (as obtained by running mstate::expand.covs).}

\item{which_group}{A character vector with the same meaning as the `groups` argument of `CoxRFX` but named (with the covariate names).}

\item{min_nr_samples}{The confidence interval of any coefficient is based on a number of bootstrap samples which at least as high as this argument. See details.}

\item{output}{Determines the sort of output. See value.}

\item{...}{Further arguments to the CoxRFX function.}
}
\value{
For each regression coefficient, the confidence intervals and the number of bootstrap samples on which they are based, if the `output` argument is equal to `CIs`; if `output` is equal to `CIs_and_coxrfx_fits`, also the CoxRFX objects for each bootstrap sample.
}
\description{
This function computes non-parametric bootstrap confidence intervals for the regression coefficients estimated by CoxRFX.
}
\author{
rc28
}
