% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convolute_Markov}
\alias{convolute_Markov}
\title{Convolution function for homogeneous Markov models}
\usage{
convolute_Markov(
  time_vector,
  diff_vector,
  probtrans_vector_1,
  probtrans_vector_2
)
}
\arguments{
\item{time_vector, integrand_1, integrand_2}{Numeric vectors.}
}
\description{
Convolution function for homogeneous Markov models
}
