% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxRFX.R
\name{CoxRFX}
\alias{CoxRFX}
\title{Empirical Bayes, multi-state Cox model}
\usage{
CoxRFX(
  Z,
  surv,
  groups = rep(1, ncol(Z)),
  which.mu = unique(groups),
  tol = 0.001,
  max.iter = 50,
  sigma0 = 0.1,
  sigma.hat = c("df", "MLE", "REML", "BLUP"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{Z}{A data frame consisting of the covariate columns of a data set in 'long format'.}

\item{surv}{A `survival' object created with \code{survival::Surv}.}

\item{groups}{A character or numeric vector whose \eqn{i}th element gives the group of the regression
coefficient associated with the \eqn{i}th covariate column of Z (coefficients belonging to the same group 
share the same Gaussian prior).}

\item{which.mu}{A vector with names or numbers of coefficient groups (see 
argument \code{groups}). If the name or number of a group of coefficients is
given in this argument, \code{CoxRFX} will estimate the mean of its Gaussian distribution;
otherwise the mean will be fixed at zero.}

\item{tol}{Convergence criterium of the EM algorithm. The algorithm stops unless
there is at least one parameter (or hyperparameter) for which it holds that the
current estimate differs in absolute terms by more than \code{tol} from the
previous estimate.}

\item{max.iter}{The maximum number of iterations in the EM algorithm.}

\item{sigma0}{A vector with the initial value of the variance hyperparameter for each group of coefficients.
Or a single value, in case the initial value of the variance hyperparameter is meant to be the same for all groups.}

\item{sigma.hat}{Which estimator to use for the variance hyperparameters (see details).}

\item{verbose}{Gives more output.}

\item{...}{Further arguments passed to the function \code{survival::coxph}.}
}
\value{
A coxph object (see \code{survival::coxph.object}) with a few extra fields: the inputs $groups, $Z, and $surv;
and the hyperparameters $sigma2 (variances) and $mu (means).
}
\description{
This function estimates a multi-state Cox model with one or more Gaussian priors
imposed on the regression coefficients (see Therneau et al., 2003).
Multiple groups of coefficients can be defined: coefficients within a group share 
the same (possibly unknown) mean and variance. The parameters and hyperparameters are
efficiently estimated by an EM-type algorithm.
}
\details{
The argument \code{Z} must be of class \code{c(data.frame,msdata)}. 

Different estimators exist for the variance hyperparameters: the default is "df", as used by Perperoglou (2014) and introduced by Schall (1991). 
Alternatives are MLE, REML, and BLUP, as defined by Therneau et al. (2003). 
Simulations suggest that the 'df' method is the most accurate.

The model can also be fitted using package \code{coxme}; the \code{coxme}
routine numerically optimises the integrated partial likelihood, which may
be more accurate, but is computationally expensive.
}
\references{
Terry M Therneau, Patricia M Grambsch & V. Shane Pankratz (2003) Penalized Survival Models and Frailty, Journal of Computational and Graphical Statistics, 12:1, 156-175, http://dx.doi.org/10.1198/1061860031365

A. Perperoglou (2014). Cox models with dynamic ridge penalties on time-varying effects of the covariates. Stat Med, 33:170-80. http://dx.doi.org/10.1002/sim.5921

R. Schall (1991). Estimation in generalized linear models with random effects. Biometrika, 78:719-727. http://dx.doi.org/10.1093/biomet/78.4.719
}
\seealso{
\code{\link{coxph.object}}; package \code{coxme}; \code{\link{Surv}}.
}
\author{
Moritz Gerstung & Rui Costa
}
