% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_auxiliary_functions.R
\name{cumhaz_splines}
\alias{cumhaz_splines}
\title{Create a spline approximation for each vector of cumulative hazards}
\usage{
cumhaz_splines(cumhaz)
}
\value{
A list with a spline function for each vector of cumulative hazards
}
\description{
Create a spline approximation for each vector of cumulative hazards
}
\author{
rc28
}
