% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebsurv-package.R
\docType{package}
\name{ebsurv-package}
\alias{ebsurv-package}
\title{Empirical Bayes multi-state Cox model}
\description{
This package implements an empirical Bayes,
multi-state Cox model.Different groups of regression
coefficients can be defined, with coefficients of the
same group sharing the same Gaussian prior. It takes
as input a data set in 'long format' and generates
estimates of relative hazards, cumulative hazard
functions and transition probabilities.
}
\author{
Moritz Gerstung, European Bioinformatics Institute EMBL-EBI
}
\keyword{package}
