% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_CIs.R
\name{extract_function}
\alias{extract_function}
\title{Ancillary function of \code{boot_probtrans}.}
\usage{
extract_function(list_object, item)
}
\value{
Bootstrap transition probabilities for target state `item`.
}
\description{
Returns all the bootstrap transition probabilities for target state `item` from a list with bootstrap transition probabilities for multiple states.
This function is not meant to be called by the user.
}
\author{
rc28
}
