% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_auxiliary_functions.R
\name{joint_cum_hazard_function}
\alias{joint_cum_hazard_function}
\title{Create a function that returns the cumulative hazard of leaving a given state}
\usage{
joint_cum_hazard_function(t, competing_transitions, spline_list)
}
\value{
A vector with the cumulative hazard of leaving a given state evaluated at given time points.
}
\description{
Create a function that returns the cumulative hazard of leaving a given state
}
\author{
rc28
}
