% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfit.R
\name{msfit}
\alias{msfit}
\alias{msfit.coxrfx}
\alias{msfit.default}
\title{Compute subject-specific transition hazards}
\usage{
msfit(object, ...)

\method{msfit}{coxrfx}(object, newdata, trans)

\method{msfit}{default}(object, newdata, variance = TRUE,
  vartype = c("aalen", "greenwood"), trans)
}
\arguments{
\item{object}{An object describing the fit of
a multi-state Cox model.}

\item{...}{other arguments}
}
\value{
An object of class "msfit"

An object of class "msfit"
}
\description{
This function computes subject-specific or overall cumulative transition
hazards for each of the possible transitions in the multi-state model.
}
