% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfit.R
\name{msfit_generic.coxrfx}
\alias{msfit_generic.coxrfx}
\title{Compute subject-specific transition hazards
under a random effects Cox model}
\usage{
\method{msfit_generic}{coxrfx}(object, newdata, trans)
}
\description{
Compute subject-specific transition hazards
under a random effects Cox model
}
