% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfit.R
\name{msfit_generic.default}
\alias{msfit_generic.default}
\title{Compute subject-specific transition hazards
under a fixed effects Cox model}
\usage{
\method{msfit_generic}{default}(
  object,
  newdata,
  variance = TRUE,
  vartype = c("aalen", "greenwood"),
  trans
)
}
\description{
Compute subject-specific transition hazards
under a fixed effects Cox model
}
