% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxRFX.R
\name{PartialC}
\alias{PartialC}
\title{Compute concordance for risk components}
\usage{
PartialC(fit, newZ = fit$Z, newSurv = fit$surv, groups = fit$groups)
}
\arguments{
\item{fit}{The CoxRFX fit}

\item{newZ}{New data, defaults to fit$Z}

\item{newSurv}{The survival object, defaults to fit$surv}

\item{groups}{The groups, defaults to fit$groups}
}
\value{
A vector with the concordance of each component
}
\description{
Compute concordance for risk components
}
\author{
mg14
}
