% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxCPSS.R
\name{print.CoxCPSS}
\alias{print.CoxCPSS}
\title{Print a CoxCPSS model}
\usage{
\method{print}{CoxCPSS}(x, ...)
}
\arguments{
\item{x}{The CoxCPSS model}

\item{...}{Currently unused}
}
\description{
Print a CoxCPSS model
}
\author{
mg14
}
