% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxRFX.R
\name{print.CoxRFX}
\alias{print.CoxRFX}
\title{Print method for CoxRFX}
\usage{
\method{print}{CoxRFX}(x, ...)
}
\arguments{
\item{x}{CoxRFX}

\item{...}{Currently unused}
}
\description{
This function implicitly calls summary.CoxRFX().
}
\author{
mg14
}
