% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_CIs.R
\name{probtrans_boot_CIs_for_target_state}
\alias{probtrans_boot_CIs_for_target_state}
\title{Ancillary function of \code{boot_probtrans}.}
\usage{
probtrans_boot_CIs_for_target_state(target_state)
}
\value{
95\% highest density confidence intervals.
}
\description{
Computes 95\% highest density bootstrap confidence 
intervals for the transition probabilities to \code{target_state}. This 
function is not meant to be called by the user.
}
\author{
rc28
}
