% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_ebsurv.R
\name{probtrans_by_convolution}
\alias{probtrans_by_convolution}
\title{Compute all transition probabilities from a given state.}
\usage{
probtrans_by_convolution(tmat, cumhaz, from_state, model)
}
\arguments{
\item{tmat}{Transition matrix.}

\item{cumhaz}{'msfit' object.}

\item{from_state}{Initial state.}

\item{model}{}
}
\description{
Compute all transition probabilities from a given state.
}
