% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_ebsurv.R
\name{probtrans_by_convolution_semiMarkov}
\alias{probtrans_by_convolution_semiMarkov}
\title{Compute transition probabilities for a given starting state and target state.}
\usage{
probtrans_by_convolution_semiMarkov(tmat, cumhaz, from_state, to_state,
  spline_list, unique_paths_object, time)
}
\arguments{
\item{tmat}{Transition matrix.}

\item{cumhaz}{'msfit' object.}

\item{from_state}{Initial state.}

\item{to_state}{Target state.}

\item{time}{}
}
\description{
Compute transition probabilities for a given starting state and target state
under a homogeneous semi-Markov model
}
