% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_ebsurv.R
\name{probtrans_ebsurv}
\alias{probtrans_ebsurv}
\title{Compute subject-specific transition probabilities
using convolution.}
\usage{
probtrans_ebsurv(initial_state, cumhaz, model)
}
\arguments{
\item{cumhaz}{An 'msfit' object}

\item{model}{Either 'Markov' or 'semiMarkov'}
}
\value{
An object of class 'probtrans'
}
\description{
Compute subject-specific transition probabilities
using convolution.
}
