% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans.R
\name{probtrans_mstate.coxrfx}
\alias{probtrans_mstate.coxrfx}
\title{Compute subject-specific or overall transition
probabilities under a fixed effects Cox model}
\usage{
\method{probtrans_mstate}{coxrfx}(object, predt, direction = c("forward", "fixedhorizon"))
}
\description{
Compute subject-specific or overall transition
probabilities under a fixed effects Cox model
}
