% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans.R
\name{probtrans_mstate.default}
\alias{probtrans_mstate.default}
\title{Compute subject-specific or overall transition
probabilities under a random effects Cox model}
\usage{
\method{probtrans_mstate}{default}(
  object,
  predt,
  direction = c("forward", "fixedhorizon"),
  method = c("aalen", "greenwood"),
  variance = TRUE,
  covariance = FALSE,
  ...
)
}
\description{
Compute subject-specific or overall transition
probabilities under a random effects Cox model
}
