% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans.R
\name{probtrans_mstate}
\alias{probtrans_mstate}
\title{Compute subject-specific or overall transition probabilities}
\usage{
probtrans_mstate(object, ...)
}
\arguments{
\item{object}{An object describing the fit of
a multi-state Cox model.}

\item{...}{other arguments}
}
\value{
An object of class "msfit"
}
\description{
This function computes subject-specific or overall transition
probabilities in multi-state models.
}
