% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxRFX.R
\name{StandardizeMagnitude}
\alias{StandardizeMagnitude}
\title{Standardize the magnitude of covariates}
\usage{
StandardizeMagnitude(X)
}
\arguments{
\item{X}{A matrix or data.frame}
}
\value{
data.frame of dim(X)
}
\description{
Standardize the magnitude of covariates
}
\author{
mg14
}
