% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_auxiliary_functions.R
\name{successful_transitions}
\alias{successful_transitions}
\title{Find the unique sequence of states between two states}
\usage{
successful_transitions(unique_paths_object, to_state)
}
\value{
A vector with the unique sequence of states between two states.
}
\description{
Find the unique sequence of states between two states
}
\author{
rc28
}
