% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxCPSS.R
\name{terms.CoxCPSS}
\alias{terms.CoxCPSS}
\title{Extract terms}
\usage{
\method{terms}{CoxCPSS}(x, ...)
}
\arguments{
\item{x}{A CoxCPSS fit}

\item{...}{Currently not used}
}
\value{
terms
}
\description{
Extract terms
}
\author{
mg14
}
