% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_auxiliary_functions.R
\name{unique_paths}
\alias{unique_paths}
\title{Find the all possible paths until absorption from a given starting state}
\usage{
unique_paths(from_state, tmat)
}
\value{
A matrix where each column is a sequence of states taken by the process until absorption. 
There are as many columns as the number of possible paths until absorption.
}
\description{
Find the all possible paths until absorption from a given starting state
}
\author{
rc28
}
