% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convolute_semiMarkov}
\alias{convolute_semiMarkov}
\title{Convolute two vectors}
\usage{
convolute_semiMarkov(time_vector, integrand_1, integrand_2)
}
\arguments{
\item{time_vector, integrand_1, integrand_2}{Numeric vectors.}
}
\description{
Convolute two vectors
}
