% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_ebsurv.R
\name{cumhaz_splines}
\alias{cumhaz_splines}
\title{Spline approximations of the cumulative hazards functions}
\usage{
cumhaz_splines(cumhaz)
}
\value{
A list with a spline function for each vector of cumulative hazards
}
\description{
Creates a spline approximation for each vector of cumulative hazards.
}
\details{
This function is not meant to be called by the user.
}
\author{
rc28
}
