% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_ebsurv.R
\name{cumhaz_splines}
\alias{cumhaz_splines}
\title{Spline approximations of the cumulative hazards functions}
\usage{
cumhaz_splines(cumhaz)
}
\value{
A list of estimated cumulative hazard functions (one for each transition).
}
\description{
Creates a spline approximation for the vector of cumulative hazards of each transition.
}
\details{
This function is used by the function 'probtrans_by_convolution'. It is not meant to be called by the user.
}
\author{
rc28
}
