% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfit.R
\name{msfit_generic}
\alias{msfit_generic}
\title{Compute subject-specific transition hazards}
\usage{
msfit_generic(object, ...)
}
\arguments{
\item{object}{An object describing the fit of
a multi-state Cox model.}

\item{...}{other arguments}
}
\value{
An object of class "msfit"
}
\description{
This function computes subject-specific or overall cumulative transition
hazards for each of the possible transitions in the multi-state model.
}
