% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_ebsurv.R
\name{successful_transitions}
\alias{successful_transitions}
\title{Find the unique sequence of states between two states}
\usage{
successful_transitions(unique_paths_object, to_state)
}
\value{
A vector with the unique sequence of states between two states.
}
\description{
\code{successful_transitions} finds the unique path between 
two states (if there is one) when the process has a tree-like structure.
}
\details{
This function is not meant to be called by the user.
}
\author{
rc28
}
