% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probtrans_auxiliary_functions.R
\name{unique_paths}
\alias{unique_paths}
\title{Find all possible paths until absorption from a given starting state}
\usage{
unique_paths(from_state, tmat)
}
\value{
A matrix where each column is a sequence of states taken by the process until absorption. 
There are as many columns as the number of possible paths until absorption.
}
\description{
\code{unique_paths} finds all possible sequences of states until absorption
when the process has tree-like structure.
}
\details{
This function is not meant to be called by the user.
}
\author{
rc28
}
